only u,d,b left

+++ spring-boot-person-service/src/main/java/microservicekerjabagus/springbootpersonservice/repository/PersonRepository.java
@@ -3,5 +3,5 @@ package microservicekerjabagus.springbootpersonservice.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import microservicekerjabagus.springbootpersonservice.entity.Person;
 
-public interface PersonRepository extends JpaRepository<Person, Long>{
+public interface PersonRepository extends JpaRepository<Person, Integer>{
 }
